# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API сервис",
        "description": "Описание API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер"
        }
    ],
    "paths": {
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Список зданий",
                "operationId": "572e55bad0698264e29e3722676f8d00",
                "responses": {
                    "200": {
                        "description": "Найденные здания",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/building/{buildingId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по зданию",
                "operationId": "991bb66e5cbbff50bd11d6c03efe0772",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по зданию",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "building": {
                                            "$ref": "#/components/schemas/Building"
                                        },
                                        "organizations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/activity/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по активности",
                "operationId": "45baece2db33adf69211e31798c924dc",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID активности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по активности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activity": {
                                            "$ref": "#/components/schemas/Activity"
                                        },
                                        "organizations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/radius": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по радиусу",
                "operationId": "a21d7085cb6e33f2ba3ee65c99585795",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организации в радиусе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/rectangle": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации в прямоугольнике",
                "operationId": "ba6e1ecff517407ccf4404d16d20c246",
                "parameters": [
                    {
                        "name": "lat_min",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lat_max",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lng_min",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lng_max",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организации в прямоугольнике",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "11826ee077ae441c61f2bee8476e2a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search/activity": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по имени активности",
                "operationId": "fd9f65d6246e92e8c9ee8849fc26492b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "activity_name"
                                ],
                                "properties": {
                                    "activity_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Организации по активности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "root_activity": {
                                            "$ref": "#/components/schemas/Activity"
                                        },
                                        "organizations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search/name/{name}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по имени",
                "operationId": "0acbc37e5194d06ca0315d65f3f0126b",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Найденные организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Organization": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationPhone"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
            "Building": {
                "required": [
                    "id",
                    "address"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Пушкина, д. 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
            "Activity": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "name": {
                        "type": "string",
                        "example": "Медицинские услуги"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrganizationPhone": {
                "required": [
                    "organization_id",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "phone": {
                        "type": "string",
                        "example": "+7 123 456 7890"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Введите ваш API-ключ",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Buildings",
            "description": "API для зданий"
        },
        {
            "name": "Organizations",
            "description": "API для организаций"
        }
    ]
}